chore: remove unused code
I did a sweep for unused code using the same method used in #3215. Most of it was leftover code that should have been removed in the recent VX HMPB interpreter purge, but there's a handful of other things that needed to be removed too.

I removed the duplicate zip helpers because `readJsonEntry` from `libs/utils` was flagged as unused, but there were two duplicates in various tests. So I removed the duplicates and used the one in `utils`.

In removing the `Interpreter` class from `apps/scan` the file no longer made sense as `interpreter.ts`, so I renamed it.
