+++ libs/cvr-fixture-generator/src/index.ts
@@ -1,2 +1,1 @@
 export * from './generate_cvrs';
-export * as GenerateCli from './cli/generate';
